+++ packages/cedar-amcharts/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- added CORS support to docs site
+- install amcharts from npm instead of git
+
 ## 1.0.0-alpha.7
 ### Changed
 - Horizontal bar charts renamed from horizontal to bar-horizontal